add basic gql queries
